Access Denied Handling Defaults
This commit introduces the capacity for users to wire denial handling
by request matcher, similar to how users can do the same with
authentication entry points. It follows the same pattern and is useful
for Resource Server when it needs to specify the access denied
response only when the request is a bearer token request.

Also note that it's arguable that this logic could be hidden for now
in the BearerTokenAccessDeniedHandler, however it felt like the code
was performing delegation as well as handling the WWW-Authenticate
response. Further, it made the ExceptionHandlingConfigure expose this
and entry points differently, which felt a bit crufty.