Improve Error Handling for Access Denied
Almost all the time, when a bearer token is denied access, it will be
due to insufficient scope.

So, for this first pass, we'll assume that it is due to insufficient
scope and report it accordingly.

In the future, we may be able to ascertain the reason in greater
detail.

Issue: gh-12