Expose a Builder
Here are some of my thoughts about what a JwtBuilder might look like
that would be simple to use in a testing framework, and may also be
useful down the road for non-testing use cases.

The main thing that I like about this approach are that it feels
similar to the existing support for a mock user authentication.