Try with inner classes
This doesn't work as-is, but I'm keeping it around to see if
something comes of it. As it is, the method binding is decided
at compile time, basic methods like XMLObjectProviderRegistry#getParserPool
fail between OpenSAML 4 and 5 since the return type's package
name chanegs.

A next thing to possibly try is to shade the opensaml-core-api
dependency for 5 in order to make the packages different.