Polish Copy Construction
The original route seemed simple in the beginning, but has since proven to be
too large of a change for such a simple request. This commit takes the approach
of using the Builder as a way to create a copy without changing the internals
of either DefaultOidcUser or DefaultOAuth2User.