Add LinkingRelyingPartyRegistrationRepository
This repository allows for accessing the relying party details in a different way from
the asserting party details. Relying party details may be stored in a database while
asserting party details are repeatedly refreshed from HTTP endpoints. This class comes
with a mechanism by which these can be queried separately by registrationId or
by a mapping from registrationId to a different value, like EntityID. It also allows the
inverse, that they be looked up by the asserting party's EntityID and then the registration
id is derived from the result.