Different Approach
Okay, I'm taking a completely different approach.

We'll see if we can wrap an AuthoritiesGranter
and an AuthenticationManager in a wrapper that implements
both AuthenticationManager and AuthorizationManager.

My thought is that this component would be built by
a configurer in the HttpSecurity DSL and would wrap the manager
used by a given authentication configurer.

This same component would then be used to wire an
AuthorizationFilter that stands in front of the
AuthenticationFilters.