Revisions
- Went back to wrapping authorization manager to grant
authorities prescribed in the DSL. I like how this keeps
the filters as they are. I also continue to imagine that
an authentication manager should be the one granting authorities
so that when an Authentication leaves the service layer,
it's immutable.

- Renamed components. AuthorizationRequestEntry is now
AuthorizationEntryPoint. I believe this is much clearer as
to its usage. Also added AuthorizationRequest to abstract
away the elements needed to determine whether a given
AuthorizationEntryPoint applies. In this way, a coder
can pass additional data, similar to how one does this
with Authentication

- Applied to more configurers and filters. Not that
the application for WebAuthn is a little odd since it
does not specify an entry poin

- Changed so that authorization entry point declarations
do not take precedence over authentication entry points
specified in the DSL. This was originally in place to
ensure that X509 was enforced. I think this is a problem
that should not be solved in this PR; we should think
about ways other than request matching to determine
which authentication entry point to use.

- Note that the above is a minor point anyway since most
other auth mechs that would normally conflict have custom
pages that sit behind AuthorizationFilter anyway.

- I'm not jazzed about the amount of boilerplate still for
specifying granting strategies. Whether it wraps or configures
an AuthenticationManager or it stands on its own in filters,
there is the matter of the default authority to grant,
any custom authorities, and the implicit granter that
updates the authentication with any authorities granted
in previous auth mechs. I'd like to simplify this more
if possible.